Split component and hook from view component
