+++ frontend/src/metabase/query_builder/components/view/sidebars/SummarizeSidebar/SummarizeContent/index.ts
@@ -1,0 +1,2 @@
+export * from "./use-summarize-query";
+export * from "./SummarizeContent";
